+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- ICS to JSON converter: events removed from original calendar are still present (see [issue 14](https://github.com/KNowledgeOnWebScale/solid-calendar-store/issues/14))
+
 ## [0.0.16] - 2022-02-01
 
 ### Fixed